re move api files to root dir
