--- .scripts/npm_start.js
@@ -1,3 +1,0 @@
-'use strict';
-
-require('@sensebox/opensensemap-api');
