+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# @sensebox/opensensemap-api Changelog
+# openSenseMap API Changelog
 
 ## Unreleased
 - Initial Release after splitting api and models
