--- api/package.json
@@ -1,60 +1,0 @@
-{
-  "name": "@sensebox/opensensemap-api",
-  "description": "openSenseMap HTTP REST API",
-  "version": "0.0.1",
-  "main": "app.js",
-  "license": "MIT",
-  "dependencies": {
-    "@turf/area": "^4.6.0",
-    "@turf/bbox": "^4.6.0",
-    "@turf/centroid": "^4.6.1",
-    "@turf/distance": "^4.6.0",
-    "@turf/helpers": "^4.6.0",
-    "@turf/hex-grid": "^4.6.1",
-    "@turf/square-grid": "^4.6.0",
-    "@turf/triangle-grid": "^4.6.0",
-    "apicache": "^0.10.0",
-    "bunyan": "^1.8.12",
-    "csv-stringify": "^1.0.4",
-    "honeybadger": "^1.2.1",
-    "isemail": "^3.0.0",
-    "millify": "^2.0.1",
-    "moment": "^2.18.1",
-    "passport": "^0.4.0",
-    "passport-jwt": "^3.0.0",
-    "passport-local": "^1.0.0",
-    "request": "^2.81.0",
-    "restify": "^5.2.0",
-    "restify-errors": "^5.0.0"
-  },
-  "devDependencies": {
-    "@sensebox/opensensemap-api-models": "file:../models",
-    "eslint": "^4.6.1"
-  },
-  "scripts": {
-    "lint": "./node_modules/.bin/eslint --ignore-pattern node_modules --fix app.js \"src/**/*.js\"",
-    "deploy-docs": "./.scripts/deploy-docs.sh",
-    "version": "node ../.scripts/npm_version-update_changelog.js && git add CHANGELOG.md"
-  },
-  "apidoc": {
-    "name": "openSenseMap API documentation",
-    "description": "methods to manage senseBoxes and get/post measurements",
-    "title": "openSenseMap API documentation",
-    "url": "https://api.opensensemap.org",
-    "order": [
-      "getBox",
-      "getBoxes",
-      "postNewBox",
-      "updateBox",
-      "deleteBox",
-      "getScript",
-      "postNewMeasurement",
-      "getMeasurements",
-      "getData",
-      "calculateIdw"
-    ]
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
