+++ lib/helpers/authHelpers.js
