+++ lib/helpers/errorHandler.js
