+++ lib/helpers/userParamHelpers.js
