+++ lib/log.js
