+++ license.md
@@ -1,6 +1,6 @@
 The MIT License (MIT)
 
-Copyright (c) 2015-2017 Matthias Pfeil, Gerald Pape
+Copyright (c) 2015 - now Matthias Pfeil
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
