+++ package.json
@@ -1,17 +1,45 @@
 {
-  "name": "opensensemap-api",
-  "private": true,
-  "version": "0.0.1",
-  "description": "API for opensensemap.org",
+  "name": "@sensebox/opensensemap-api",
+  "version": "0.0.0",
+  "description": "openSenseMap HTTP REST API",
+  "main": "app.js",
+  "license": "MIT",
+  "author": "Matthias Pfeil",
+  "contributors": [
+    "Christoph Kisfeld",
+    "Gerald Pape",
+    "Norwin Roosen"
+  ],
   "dependencies": {
-    "@sensebox/opensensemap-api": "file:./api",
-    "@sensebox/opensensemap-api-models": "file:./models"
+    "@sensebox/opensensemap-api-models": "file:./models",
+    "@turf/area": "^4.6.0",
+    "@turf/bbox": "^4.6.0",
+    "@turf/centroid": "^4.6.1",
+    "@turf/distance": "^4.6.0",
+    "@turf/helpers": "^4.6.0",
+    "@turf/hex-grid": "^4.6.1",
+    "@turf/square-grid": "^4.6.0",
+    "@turf/triangle-grid": "^4.6.0",
+    "apicache": "^0.10.0",
+    "bunyan": "^1.8.12",
+    "csv-stringify": "^1.0.4",
+    "honeybadger": "^1.2.1",
+    "isemail": "^3.0.0",
+    "millify": "^2.0.1",
+    "moment": "^2.18.1",
+    "passport": "^0.4.0",
+    "passport-jwt": "^3.0.0",
+    "passport-local": "^1.0.0",
+    "request": "^2.81.0",
+    "restify": "^5.2.0",
+    "restify-errors": "^5.0.0"
   },
   "devDependencies": {
     "@turf/invariant": "^4.6.0",
     "chakram": "^1.5.0",
     "cheerio": "^1.0.0-rc.2",
     "csv-parse": "^1.2.1",
+    "eslint": "^4.6.1",
     "mimelib": "^0.3.1",
     "mocha": "^3.5.0",
     "moment": "^2.18.1",
@@ -19,19 +47,43 @@
     "request-promise-native": "^1.0.4"
   },
   "scripts": {
-    "start": "node ./.scripts/npm_start.js",
+    "start": "node app.js",
     "test": "node node_modules/.bin/mocha tests/tests.js",
     "pretest": "node tests/waitForHttp",
-    "tag-container": "./.scripts/npm_tag-container.sh"
+    "tag-container": "./.scripts/npm_tag-container.sh",
+    "lint": "./node_modules/.bin/eslint --ignore-pattern node_modules --fix app.js \"src/**/*.js\"",
+    "deploy-docs": "./.scripts/deploy-docs.sh"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/sensebox/openSenseMap-API.git"
   },
-  "author": "Matthias Pfeil",
-  "license": "MIT",
   "bugs": {
     "url": "https://github.com/sensebox/openSenseMap-API/issues"
   },
-  "homepage": "https://github.com/sensebox/openSenseMap-API"
+  "homepage": "https://github.com/sensebox/openSenseMap-API",
+  "apidoc": {
+    "name": "openSenseMap API documentation",
+    "description": "methods to manage senseBoxes and get/post measurements",
+    "title": "openSenseMap API documentation",
+    "url": "https://api.opensensemap.org",
+    "order": [
+      "getBox",
+      "getBoxes",
+      "postNewBox",
+      "updateBox",
+      "deleteBox",
+      "getScript",
+      "postNewMeasurement",
+      "getMeasurements",
+      "getData",
+      "calculateIdw"
+    ],
+    "template": {
+      "withCompare": false
+    }
+  },
+  "publishConfig": {
+    "access": "public"
+  }
 }
